Add GitHub Actions workflow for publishing
Added a new GitHub Actions workflow configuration file named `publish.yml`.
The workflow is named `publish` and can be triggered manually from the GitHub UI, on push events to the `master` branch, on all pull requests, and when a new GitHub release is published.

Set environment variables to skip the first-time experience for .NET and suppress the logo, and defined a directory for NuGet packages. Configured the default shell to be PowerShell (`pwsh`).

Defined four jobs within the workflow:
1. `create_nuget`:
   - Runs on `ubuntu-latest`.
   - Checks out the repository with full history.
   - Sets up the .NET SDK.
   - Creates a NuGet package in the specified directory.
   - Uploads the created NuGet package as an artifact.
2. `validate_nuget`:
   - Runs on `ubuntu-latest`.
   - Depends on the `create_nuget` job.
   - Sets up the .NET SDK.
   - Downloads the NuGet package artifact.
   - Installs the `Meziantou.Framework.NuGetPackageValidation.Tool`.
   - Validates the NuGet package using the installed tool.
3. `run_test`:
   - Runs on `ubuntu-latest`.
   - Checks out the repository.
   - Sets up the .NET SDK.
   - Runs the tests in Release configuration.
4. `deploy`:
   - Runs on `ubuntu-latest`.
   - Depends on the `validate_nuget` and `run_test` jobs.
   - Only runs when a GitHub release is created.
   - Downloads the NuGet package artifact.
   - Sets up the .NET SDK.
   - Publishes the NuGet packages to NuGet.org, using the `--skip-duplicate` flag to avoid errors if the package already exists.